Create Commutative Properties of the Dot Product.py
This code performs tests to check the commutativity of dot products. It generates random vectors and computes a^T * b and b^T * a to test for equality. It also tests with fixed 2-element vectors and 1000-element random vectors. The code counts the number of times the dot products are not equal to each other.